few changes
there were no getters and setters for the new instance variables added
to the children classes so i had to add those

there were calls in usecars to arg-less constructors for sedans and
sports but neither classes have arg-less constructors... now this is
what i think i learned about this.. to make sure there's never a car
without a carowner, i don't have any constructors that don't have least
an arg to pass in a carowner. but what i did just to make your code
compile was add the arg-less constructors, but i wonder what you think
about having arg-less when there's suppose to be composition... with an
argless constructor, it would be possible to instantiate that object
type without having an object of the type that is needed for that
dependent object. if that makes any sense....

now the code compiles